+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # COMING NEXT!
 
 * DDB Icon image selection was not always respected.
+* Some monsters would cause the use SRD images copy to fail.
 
 # 3.3.15